tests/wpt/mozilla/tests/mozilla/bluetooth/getPrimaryServices/blacklisted-services.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.blacklist);
    return window.navigator.bluetooth.requestDevice({filters: [{services: [device_information.name]}], optionalServices: [human_interface_device.name]})
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryServices())
    .then(services => {
        assert_equals(services.length, 4);
        assert_equals(services[0].uuid,
                      BluetoothUUID.getService(blacklist_test_service_uuid));
        assert_equals(services[1].uuid,
                      BluetoothUUID.getService('device_information'));
        assert_equals(services[2].uuid,
                      BluetoothUUID.getService('generic_access'));
        assert_equals(services[3].uuid,
                      BluetoothUUID.getService('heart_rate'));
    });
}, 'Request for services. Does not return blacklisted service.');
</script>
